refactor(85-refactor-services): move response / dto classes to dto package
