+++ server/src/main/java/fr/bookhub/controller/BookController.java
@@ -1,13 +1,14 @@
 package fr.bookhub.controller;
 
+import fr.bookhub.dto.BookCreateRequest;
+import fr.bookhub.dto.BookResponse;
 import fr.bookhub.service.*;
 import fr.bookhub.service.filter.BookSearchFilter;
+import fr.bookhub.utility.MethodType;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @AllArgsConstructor
