+++ server/src/main/java/fr/bookhub/dto/AuthorCreateRequest.java
@@ -1,6 +1,5 @@
-package fr.bookhub.service;
+package fr.bookhub.dto;
 
-import fr.bookhub.entity.Role;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,10 +9,8 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserBasicResponse {
-    private Integer id;
-    private String lastName;
+public class AuthorCreateRequest {
     private String firstName;
-    private String pseudo;
-    private Role role;
+    private String lastName;
+    private String country;
 }
