+++ server/src/main/java/fr/bookhub/dto/LoanResponse.java
@@ -1,10 +1,6 @@
-package fr.bookhub.service;
+package fr.bookhub.dto;
 
-import fr.bookhub.entity.Book;
 import fr.bookhub.entity.Status;
-import fr.bookhub.entity.User;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
