+++ server/src/main/java/fr/bookhub/dto/UserBasicMapper.java
@@ -1,4 +1,4 @@
-package fr.bookhub.service;
+package fr.bookhub.dto;
 
 import fr.bookhub.entity.User;
 import org.springframework.stereotype.Component;
